modules/location_consumer_service/requirements.txt
SQLAlchemy==1.3.19
python-dotenv==0.19.0
kafka-python==2.0.2
psycopg2-binary==2.8.5